requirements.txt
pytorch==1.12.1
tensorboard==2.10.1
torchvision==0.13.1
configargparse==1.4
numpy==1.21.5
scikit_learn==1.1.2
timm==0.5.4 